requirements.txt
fastapi
pydantic-settings
uvicorn[standard]
sqlalchemy
databases[aiosqlite]
python-dotenv
rich
asgi-correlation-id
python-json-logger
logtail-python
python-jose
python-multipart
passlib[bcrypt]
httpx
werkzeug
minio